[Refactor] #25 메시지와 채팅 참여 패키지를 채팅 서브 패키지로 이동 (#26)
* refactor: 메시지 패키지를 채팅 서브 패키지로 이동

* refactor: 채팅 참여 패키지를 채팅 서브 패키지로 이동
