--- src/main/java/com/lovesoongalarm/lovesoongalarm/domain/chat/sub/message/persistence/repository/MessageRepository.java
@@ -1,7 +1,0 @@
-package com.lovesoongalarm.lovesoongalarm.domain.chat.sub.message.persistence.repository;
-
-import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.message.persistence.entity.Message;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MessageRepository extends JpaRepository<Message, Long> {
-}
