+++ src/main/java/com/lovesoongalarm/lovesoongalarm/domain/chat/sub/room/sub/participant/implement/ChatRoomParticipantSaver.java
@@ -1,7 +1,7 @@
-package com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.implement;
+package com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.sub.participant.implement;
 
-import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.persistence.entity.ChatRoomParticipant;
-import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.persistence.repository.ChatRoomParticipantRepository;
+import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.sub.participant.persistence.entity.ChatRoomParticipant;
+import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.sub.participant.persistence.repository.ChatRoomParticipantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
